Attempt to have a single debugger entry point
This was an attempt to workaround the issues in Code where we can't
correctly select the debug adapter based on what the proejct is. It
failed... The problem seemed to be that if we new our own DebugSession
up, it hasn't had all of the correct init methods called, so
this.sendEvent(etc.) do not function.

I tried passing a reference to MultiDebugSession into the specific
sessions, however things got really messy.

Pushing to a branch in case I find a way to do it (and this turns out
the best way to do it).